requirements.txt
Django>=1.11,<2
django-extensions==1.7.5
djangorestframework==3.5.3
django-crispy-forms==1.6.1
